+++ src/types/account.ts
@@ -5,5 +5,5 @@ export interface LocalAccountInstance {
 }
 
 export interface Login {
-  callbackRoute?: string;
+  callbackUrl?: string;
 }